refactor: split ProxyHostInfoPanel into focused sub-components
Part of #103
