+++ src/components/features/proxy-hosts/ProxyHostBasicInfoSection.tsx
@@ -1,0 +1,140 @@
+import {
+  Typography,
+  Box,
+  Grid,
+  IconButton,
+} from '@mui/material'
+import {
+  ContentCopy as CopyIcon,
+  Person as PersonIcon,
+  SwapHoriz as SwapHorizIcon,
+} from '@mui/icons-material'
+import { ProxyHost } from '../../../api/proxyHosts'
+import { formatDate } from '../../../utils/dateUtils'
+
+interface ProxyHostBasicInfoSectionProps {
+  host: ProxyHost
+  onCopyToClipboard: (text: string, label?: string) => void
+}
+
+const ProxyHostBasicInfoSection = ({ host, onCopyToClipboard }: ProxyHostBasicInfoSectionProps) => (
+  <>
+    <Box
+      sx={{
+        display: "flex",
+        alignItems: "center",
+        gap: 1,
+        mb: 2
+      }}>
+      <SwapHorizIcon color="primary" />
+      <Typography variant="h6">Basic Information</Typography>
+    </Box>
+
+    <Grid container spacing={2}>
+      <Grid size={{ xs: 12, md: 6 }}>
+        <Typography
+          variant="subtitle2"
+          gutterBottom
+          sx={{
+            color: "text.secondary",
+            fontWeight: "bold"
+          }}>
+          Host ID
+        </Typography>
+        <Box
+          sx={{
+            display: "flex",
+            alignItems: "center",
+            gap: 1
+          }}>
+          <Typography variant="body2" sx={{
+            fontFamily: "monospace"
+          }}>
+            #{host.id}
+          </Typography>
+          <IconButton
+            size="small"
+            aria-label="Copy to clipboard"
+            onClick={() => onCopyToClipboard(host.id.toString(), 'Host ID')}
+          >
+            <CopyIcon fontSize="small" />
+          </IconButton>
+        </Box>
+      </Grid>
+
+      <Grid size={{ xs: 12, md: 6 }}>
+        <Typography
+          variant="subtitle2"
+          gutterBottom
+          sx={{
+            color: "text.secondary",
+            fontWeight: "bold"
+          }}>
+          Forward Destination
+        </Typography>
+        <Typography variant="body2" sx={{
+          fontFamily: "monospace"
+        }}>
+          {host.forward_scheme}://{host.forward_host}:{host.forward_port}
+        </Typography>
+      </Grid>
+
+      <Grid size={{ xs: 12, md: 6 }}>
+        <Typography
+          variant="subtitle2"
+          gutterBottom
+          sx={{
+            color: "text.secondary",
+            fontWeight: "bold"
+          }}>
+          Created
+        </Typography>
+        <Typography variant="body2">
+          {formatDate(host.created_on)}
+        </Typography>
+      </Grid>
+
+      <Grid size={{ xs: 12, md: 6 }}>
+        <Typography
+          variant="subtitle2"
+          gutterBottom
+          sx={{
+            color: "text.secondary",
+            fontWeight: "bold"
+          }}>
+          Last Modified
+        </Typography>
+        <Typography variant="body2">
+          {formatDate(host.modified_on)}
+        </Typography>
+      </Grid>
+
+      {host.owner_user_id && (
+        <Grid size={{ xs: 12, md: 6 }}>
+          <Typography
+            variant="subtitle2"
+            gutterBottom
+            sx={{
+              color: "text.secondary",
+              fontWeight: "bold"
+            }}>
+            Owner
+          </Typography>
+          <Box
+            sx={{
+              display: "flex",
+              alignItems: "center",
+              gap: 0.5
+            }}>
+            <PersonIcon fontSize="small" color="action" />
+            <Typography variant="body2">
+              User #{host.owner_user_id}
+            </Typography>
+          </Box>
+        </Grid>
+      )}
+    </Grid>
+  </>
+)
+
+export default ProxyHostBasicInfoSection
